Validate email, password, number and text entries

+++ nest-e_commerce-app/apps/shop/src/shop.service.ts
@@ -3,6 +3,9 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Shop } from 'apps/shop/src/shop.entity';
 
+import { stringValidator, numberValidator } from 'apps/utils/validators';
+import { stringErrorText, numberErrorText } from 'apps/utils/text';
+
 @Injectable()
 export class ShopService {
   constructor(
@@ -15,6 +18,9 @@ export class ShopService {
   }
 
   async findOne(id: number): Promise<Shop> {
+    if (!numberValidator(id, numberErrorText)) {
+      throw new NotFoundException(numberErrorText);
+    }
     const shop = await this.shopRepository.findOneBy({ id });
     if (!shop) {
       throw new NotFoundException(`Shop with ID ${id} not found`);
@@ -24,11 +30,24 @@ export class ShopService {
 
   async create(shop: Partial<Shop>): Promise<Shop> {
     const newShop = this.shopRepository.create(shop);
+    if (
+      !stringValidator(newShop.name, stringErrorText) ||
+      !stringValidator(newShop.themeConfig, stringErrorText)
+    ) {
+      throw new NotFoundException(stringErrorText);
+    }
+    if (!numberValidator(newShop.ownerId, numberErrorText)) {
+      throw new NotFoundException(numberErrorText);
+    }
+
     console.log('newShop : ', newShop);
     return this.shopRepository.save(newShop);
   }
 
   async update(id: number, shop: Partial<Shop>): Promise<Shop> {
+    if (!numberValidator(id, numberErrorText)) {
+      throw new NotFoundException(numberErrorText);
+    }
     await this.shopRepository.update(id, shop);
     const updatedShop = await this.shopRepository.findOneBy({ id });
     if (!updatedShop) {
@@ -38,6 +57,9 @@ export class ShopService {
   }
 
   async remove(id: number): Promise<void> {
+    if (!numberValidator(id, numberErrorText)) {
+      throw new NotFoundException(numberErrorText);
+    }
     const result = await this.shopRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Shop with ID ${id} not found`);
