+++ nest-e_commerce-app/apps/users/src/users.service.ts
@@ -1,7 +1,21 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  BadRequestException,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './users.entity';
+import {
+  emailValidator,
+  passwordValidator,
+  numberValidator,
+} from '../../utils/validators';
+import {
+  passwordErrorText,
+  emailErrorText,
+  numberErrorText,
+} from 'apps/utils/text';
 
 @Injectable()
 export class UsersService {
@@ -14,17 +28,43 @@ export class UsersService {
     return this.usersRepository.find();
   }
 
-  findOne(email: string): Promise<User | null> {
-    return this.usersRepository.findOneBy({ email });
+  async findOne(email: string): Promise<User | null> {
+    const validEmail = emailValidator(email, emailErrorText);
+    if (!validEmail) {
+      throw new BadRequestException(emailErrorText);
+    }
+
+    const user = await this.usersRepository.findOneBy({ email });
+    if (!user) {
+      throw new NotFoundException(`User with email ${email} not found`);
+    }
+    return user;
   }
 
-  async create(email: string, password: string): Promise<User> {
+  async create(email: string, password: string): Promise<User | string> {
+    const validEmail = emailValidator(email, emailErrorText);
+    const validPassword = passwordValidator(password, passwordErrorText);
+
+    if (!validEmail) {
+      throw new BadRequestException(emailErrorText);
+    }
+    if (!validPassword) {
+      throw new BadRequestException(passwordErrorText);
+    }
+
     const newUser = this.usersRepository.create({ email, password });
-    console.log('newUser', newUser);
     return this.usersRepository.save(newUser);
   }
 
   async remove(id: number): Promise<void> {
-    await this.usersRepository.delete(id);
+    const validId = numberValidator(id, numberErrorText);
+    if (!validId) {
+      throw new BadRequestException(numberErrorText);
+    }
+
+    const result = await this.usersRepository.delete(id);
+    if (result.affected === 0) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
   }
 }
